test: move worker_threads dc tests to own file
worker_threads.test.ts has ~10 pre-existing 5s timeouts under the ASAN
debug build that are unrelated to this change, so the gate can never see
the full file pass. The directory already uses per-topic test files
(worker-async-dispose.test.ts, worker-transfer-list.test.ts, etc).
